requirements.txt
lgpio
rpi-ws281x
pyserial
